server/requirements.txt
alabaster==0.7.16
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
astatine==0.3.3
astor==0.8.1
astpretty==3.0.0
astroid==2.15.8
asttokens==2.4.1
attrs==23.2.0
autoflake==1.7.8
Babel==2.14.0
bandit==1.7.8
beautifulsoup4==4.12.3
bidict==0.23.1
black==23.12.1
bleach==6.1.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cognitive_complexity==1.3.0
colorama==0.4.6
coverage==7.5.0
cryptography==42.0.5
darglint==1.8.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
dlint==0.14.1
doc8==1.1.1
docformatter==1.7.5
docutils==0.19
domdf-python-tools==3.8.0.post2
eradicate==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.110.3
fastjsonschema==2.19.1
filelock==3.14.0
flake8==5.0.4
flake8-2020==1.8.1
flake8-aaa==0.17.0
flake8-annotations==3.0.1
flake8-annotations-complexity==0.0.8
flake8-annotations-coverage==0.0.6
flake8-bandit==4.1.1
flake8-black==0.3.6
flake8-blind-except==0.2.1
flake8-breakpoint==1.1.0
flake8-broken-line==0.6.0
flake8-bugbear==23.3.12
flake8-builtins==1.5.3
flake8-class-attributes-order==0.1.3
flake8-coding==1.3.2
flake8-comments==0.1.2
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-django==1.4
flake8-docstrings==1.7.0
flake8-encodings==0.5.1
flake8-eradicate==1.5.0
flake8-executable==2.1.3
flake8-expression-complexity==0.0.11
flake8-fastapi==0.7.0
flake8-fixme==1.1.1
flake8-functions==0.0.8
flake8-future-annotations==0.0.5
flake8-helper==0.2.2
flake8-isort==6.1.1
flake8-literal==1.4.0
flake8-logging-format==0.9.0
flake8-markdown==0.5.0
flake8-mutable==1.2.0
flake8-no-pep420==2.7.0
flake8-noqa==1.4.0
flake8-pie==0.16.0
flake8-plugin-utils==1.3.3
flake8-pyi==22.11.0
flake8-pylint==0.2.1
flake8-pytest-style==1.7.2
flake8-quotes==3.4.0
flake8-rst-docstrings==0.3.0
flake8-secure-coding-standard==1.4.1
flake8-string-format==0.3.0
flake8-tidy-imports==4.10.0
flake8-typing-imports==1.15.0
flake8-use-fstring==1.4
flake8-use-pathlib==0.3.0
flake8-useless-assert==0.4.4
flake8-variables-names==0.0.6
flake8-warnings==0.4.0
flake8_cognitive_complexity==0.1.0
flake8_functions_names==0.4.0
flake8_simplify==0.21.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
gcloud==0.18.3
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httplib2==0.22.0
hypothesis==6.100.2
hypothesmith==0.1.9
idna==3.7
imagesize==1.4.1
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab-flake8==0.7.1
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
lark-parser==0.12.0
lazy-object-proxy==1.10.0
libcst==0.4.10
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mr-proper==0.0.7
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
oauth2client==4.1.3
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pandas-vet==0.2.3
pandocfilters==1.5.1
pathspec==0.12.1
pbr==6.0.0
pep8-naming==0.13.3
platformdirs==4.2.1
pluggy==1.5.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybetter==0.4.1
pycln==2.4.0
pycodestyle==2.9.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydocstyle==6.3.0
pyemojify==0.2.0
pyflakes==2.5.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==2.17.7
PyMySQL==1.1.0
pyparsing==3.1.2
pyproject-api==1.6.1
pytest==7.4.4
pytest-cov==4.1.0
pytest-sugar==0.9.7
python-dateutil==2.9.0.post0
python-dev-tools==2023.3.24
python-dotenv==1.0.1
python-engineio==4.9.1
python-jwt==4.1.0
python-socketio==5.11.2
pytz==2024.1
pyupgrade==3.15.2
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
referencing==0.35.0
removestar==1.5
requests==2.32.2
requests-toolbelt==0.10.1
restructuredtext_lint==1.4.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==6.2.1
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.29
ssort==0.13.0
starlette==0.37.2
stdlib-list==0.10.0
stevedore==5.2.0
termcolor==2.4.0
tinycss2==1.3.0
tokenize-rt==5.2.0
tomli==2.0.1
tomlkit==0.12.4
tornado==6.4
tox==4.15.0
tox-travis==0.12
traitlets==5.14.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
untokenize==0.1.1
urllib3==1.26.18
virtualenv==20.26.1
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
gunicorn==20.0.4




